Update lib Webdriver Manager

+++ actions/commons/BaseTest.java
@@ -62,7 +62,7 @@ public class BaseTest {
     }
     public String getRandomEmail() {
         Random rand = new Random();
-        return "automation" + rand.nextLong(9999) + "@gmail.com";
+        return "automation" + rand.nextInt(9999) + "@gmail.com";
     }
 
     public void quitBrowser(){
