+++ testcases/com/nopcommerce/account/Level_02_Register_Static_Method.java
@@ -115,7 +115,7 @@ public class Level_02_Register_Static_Method {
 
     public String getRandomEmail() {
         Random rand = new Random();
-        return  "automation" + rand.nextLong(9999) + "@gmail.com";
+        return  "automation" + rand.nextInt(9999) + "@gmail.com";
     }
 
     @AfterClass
